+++ dlib/filesystem/stdposixdir.d
@@ -38,7 +38,7 @@ version(Posix)
 }
 import dlib.core.memory;
 import dlib.filesystem.filesystem;
-import dlib.text.unmanagedstring;
+import dlib.text.str;
 
 version(Posix):